assistant/agent/requirements.txt
google-cloud-bigquery>=3.13.0
google-adk>=0.1.0
pyyaml>=6.0
